refactor(providers): split remaining services out of app_providers.dart, close #4506
Final sub-batch (9c of 3) of #4506's split wave. Refs #4339 Wave 2.

Moves the 15-entity social/foundation/collaborator cluster to
lib/providers/social_providers.dart:

- 3 legacy Provider<X> collaborator providers (response service, invite
  state store, confirmation repository)
- pendingActionService, outgoingDmRetryService (offline-action and DM
  retry orchestration)
- analyticsService (Kind 22236 view publisher)
- hashtagCacheService, hashtagService (in-memory + persistent hashtag indexes)
- draftStorageService, clipLibraryService (per-user content drafts/clips)
- userDataCleanupService (sign-out / identity-change cleanup orchestrator)
- socialService (NIP-51 Kind 30000 follow sets)
- contentReportingService (NIP-56)
- contentDeletionService (NIP-09)
- bugReportService (NIP-17 encrypted bug reports + Blossom uploads)

Migration pattern (mirrors 9a/9b):

  // app_providers.dart
  export 'social_providers.dart';  // keeps consumers' imports unchanged

After this batch, app_providers.dart is essentially a barrel of re-exports
plus the lone createBlockedAuthorFilter helper (which still has 3 callers
across moderation, repository, and video providers — moving it is tracked
under #4554 along with the rest of the back-import cleanup).

Scope cleanup:

- dart fix --apply removed 33 newly-unused imports from app_providers.dart.
- Deleted app_providers.g.dart entirely — no @riverpod annotations remain
  in app_providers.dart, so the generated file has nothing to produce.
  Dropped the `part 'app_providers.g.dart';` directive.
- Three section dividers removed (FOUNDATIONAL SERVICES, DEPENDENT
  SERVICES, "(Removed duplicate legacy provider for StreamUploadService)")
  since their content all moved out.

Cumulative reduction:
- app_providers.dart: 2,918 -> 40 LoC (98.6% reduction).
- Domain files now own ~all provider declarations. Each provider file is
  under 800 LoC (#4506 acceptance criterion).

Closes #4506.

Test plan:
- dart run build_runner build --delete-conflicting-outputs (clean)
- flutter analyze lib test integration_test (No issues found)
- flutter test test/screens/feed/pooled_video_feed_item_repo_swap_test.dart
  test/widgets/video_feed_item/feed_videos_repo_swap_test.dart
  test/providers/dm_repository_provider_test.dart
  test/providers/curated_list_repository_bridge_test.dart (10 passed)
